backened/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.4.4
Werkzeug==2.3.6
psycopg2-binary==2.9.7
gunicorn==21.2.0
